+++ modules/core/arrow-core-data/src/test/kotlin/arrow/core/NonEmptyListTest.kt
@@ -5,6 +5,7 @@ import arrow.core.extensions.hash
 import arrow.core.extensions.nonemptylist.applicative.applicative
 import arrow.core.extensions.nonemptylist.comonad.comonad
 import arrow.core.extensions.nonemptylist.eq.eq
+import arrow.core.extensions.nonemptylist.functorFilter.functorFilter
 import arrow.core.extensions.nonemptylist.hash.hash
 import arrow.core.extensions.nonemptylist.monad.monad
 import arrow.core.extensions.nonemptylist.semigroup.semigroup
@@ -13,6 +14,7 @@ import arrow.core.extensions.nonemptylist.show.show
 import arrow.core.extensions.nonemptylist.traverse.traverse
 import arrow.test.UnitSpec
 import arrow.test.laws.ComonadLaws
+import arrow.test.laws.FunctorFilterLaws
 import arrow.test.laws.HashLaws
 import arrow.test.laws.MonadLaws
 import arrow.test.laws.SemigroupKLaws
@@ -40,11 +42,5 @@ class NonEmptyListTest : UnitSpec() {
       SemigroupLaws.laws(NonEmptyList.semigroup(), Nel(1, 2, 3), Nel(3, 4, 5), Nel(6, 7, 8), NonEmptyList.eq(Int.eq())),
       HashLaws.laws(NonEmptyList.hash(Int.hash()), EQ) { Nel.of(it) }
     )
-
-    "mapFilter" {
-      val op: NonEmptyList<Int> = NonEmptyList(0, List(100) { s: Int -> 10 * s })
-      val res = op.foldLeft(NonEmptyList(0)) { acc, i -> if (i < 44) acc else acc + i }
-      op.mapFilter { if (it < 44) None else Some(it) } == res
-    }
   }
 }
