+++ modules/core/arrow-core-data/src/test/kotlin/arrow/core/SequenceKTest.kt
@@ -3,9 +3,9 @@ package arrow.core
 import arrow.Kind
 import arrow.core.extensions.eq
 import arrow.core.extensions.hash
-import arrow.core.extensions.list.foldable.foldLeft
 import arrow.core.extensions.sequencek.applicative.applicative
 import arrow.core.extensions.sequencek.eq.eq
+import arrow.core.extensions.sequencek.functorFilter.functorFilter
 import arrow.core.extensions.sequencek.hash.hash
 import arrow.core.extensions.sequencek.monad.monad
 import arrow.core.extensions.sequencek.monoid.monoid
@@ -14,6 +14,7 @@ import arrow.core.extensions.sequencek.monoidal.monoidal
 import arrow.core.extensions.sequencek.traverse.traverse
 import arrow.test.UnitSpec
 import arrow.test.generators.sequenceK
+import arrow.test.laws.FunctorFilterLaws
 import arrow.test.laws.HashLaws
 import arrow.test.laws.MonadLaws
 import arrow.test.laws.MonoidKLaws
@@ -23,7 +24,6 @@ import arrow.test.laws.ShowLaws
 import arrow.test.laws.TraverseLaws
 import arrow.typeclasses.Eq
 import arrow.typeclasses.Show
-import io.kotlintest.matchers.sequences.shouldContainAll
 import io.kotlintest.properties.Gen
 import io.kotlintest.runner.junit4.KotlinTestRunner
 import org.junit.runner.RunWith
@@ -35,22 +35,22 @@ class SequenceKTest : UnitSpec() {
 
     val eq: Eq<Kind<ForSequenceK, Int>> = object : Eq<Kind<ForSequenceK, Int>> {
       override fun Kind<ForSequenceK, Int>.eqv(b: Kind<ForSequenceK, Int>): Boolean =
-        toList() == b.toList()
+        fix().toList() == b.fix().toList()
     }
 
     val associativeSemigroupalEq: Eq<Kind<ForSequenceK, Tuple2<Int, Tuple2<Int, Int>>>> = object : Eq<Kind<ForSequenceK, Tuple2<Int, Tuple2<Int, Int>>>> {
       override fun Kind<ForSequenceK, Tuple2<Int, Tuple2<Int, Int>>>.eqv(b: Kind<ForSequenceK, Tuple2<Int, Tuple2<Int, Int>>>): Boolean =
-        this.toList() == b.toList()
+        this.fix().toList() == b.fix().toList()
     }
 
     val tuple2Eq: Eq<Kind<ForSequenceK, Tuple2<Int, Int>>> = object : Eq<Kind<ForSequenceK, Tuple2<Int, Int>>> {
       override fun Kind<ForSequenceK, Tuple2<Int, Int>>.eqv(b: Kind<ForSequenceK, Tuple2<Int, Int>>): Boolean =
-        toList() == b.toList()
+        fix().toList() == b.fix().toList()
     }
 
     val show: Show<Kind<ForSequenceK, Int>> = object : Show<Kind<ForSequenceK, Int>> {
       override fun Kind<ForSequenceK, Int>.show(): String =
-        toList().toString()
+        fix().toList().toString()
     }
 
     testLaws(
@@ -60,14 +60,9 @@ class SequenceKTest : UnitSpec() {
       MonoidLaws.laws(SequenceK.monoid(), Gen.sequenceK(Gen.int()), eq),
       MonoidalLaws.laws(SequenceK.monoidal(), { SequenceK.just(it) }, tuple2Eq, this::bijection, associativeSemigroupalEq),
       TraverseLaws.laws(SequenceK.traverse(), SequenceK.applicative(), { n: Int -> SequenceK(sequenceOf(n)) }, eq),
+      FunctorFilterLaws.laws(SequenceK.functorFilter(), { SequenceK.just(it) }, eq),
       HashLaws.laws(SequenceK.hash(Int.hash()), SequenceK.eq(Int.eq())) { sequenceOf(it).k() }
     )
-
-    "mapFilter" {
-      val op: SequenceK<Int> = List(100) { s: Int -> 10 * s }.asSequence().k()
-      val res = List(100) { s: Int -> 10 * s }.foldLeft(emptySequence<Int>()) { acc, i -> if (i < 44) acc else acc + i }.k()
-      op.mapFilter { if (it < 44) None else Some(it) } shouldContainAll res
-    }
   }
 
   private fun bijection(from: Kind<ForSequenceK, Tuple2<Tuple2<Int, Int>, Int>>): SequenceK<Tuple2<Int, Tuple2<Int, Int>>> =
