--- modules/core/arrow-core/src/main/kotlin/arrow/core/extensions/NonEmptyListFunctorFilter.kt
@@ -1,18 +1,0 @@
-package arrow.core.extensions
-
-import arrow.Kind
-import arrow.core.Option
-import arrow.core.ForNonEmptyList
-import arrow.core.NonEmptyList
-import arrow.core.fix
-import arrow.extension
-import arrow.typeclasses.FunctorFilter
-
-@extension
-interface NonEmptyListFunctorFilter : FunctorFilter<ForNonEmptyList> {
-  override fun <A, B> Kind<ForNonEmptyList, A>.mapFilter(f: (A) -> Option<B>): NonEmptyList<B> =
-    fix().mapFilter(f)
-
-  override fun <A, B> Kind<ForNonEmptyList, A>.map(f: (A) -> B): NonEmptyList<B> =
-    fix().map(f)
-}
