--- modules/core/arrow-core/src/main/kotlin/arrow/core/extensions/TryFunctorFilter.kt
@@ -1,19 +1,0 @@
-package arrow.core.extensions
-
-import arrow.Kind
-import arrow.core.ForTry
-import arrow.core.Option
-import arrow.core.Try
-import arrow.core.fix
-import arrow.extension
-import arrow.typeclasses.FunctorFilter
-
-@extension
-interface TryFunctorFilter : FunctorFilter<ForTry> {
-
-  override fun <A, B> Kind<ForTry, A>.mapFilter(f: (A) -> Option<B>): Try<B> =
-    fix().mapFilter(f)
-
-  override fun <A, B> Kind<ForTry, A>.map(f: (A) -> B): Try<B> =
-    fix().map(f)
-}
